+++ HISTORY.md
@@ -6,6 +6,7 @@
 - add game "Twilight Zone (FreeWPC), Broken"
 - add game "Attack from Mars (FreeWPC), Broken"
 - add game "White Water "Bigfoot" (FreeWPC)"
+- add game "Demolition Man "Demolition Time" (FreeWPC)"
 - add keyboard mapping "1" - "4"
 
 ## 14/1/19